Update Install-PowerShell.ps1 add hash validation and default temp file clean up
The file SHA256 hashes appear to be provided in the release notes on the version tag.
Added a process to use a regular expression that pulls out the expected hash value 
for the file from the website and compares that to a calculated hash of the file.

The two hashes should match, and if they don't then it's likely that the file has been corrupted.

Also it's a bit of a breaking change, but cleaned up the file in the temp folder once
it has been installed. If this is not desired it can be removed, or the caller can
also use the new -KeepDownload option as well.